requirements.txt
google-generativeai
chromadb
networkx
python-dotenv
numpy
